+++ dimcli/utils/utils_dimensions.py
@@ -1,5 +1,11 @@
 """
-DSL / API specific utilities
+This module contains utilities for querying the API and working with Dimensions data.  
+
+NOTE: these functions are attached to the top level ``dimcli.utils`` module. E.g.:
+
+>>> from dimcli.utils import dimensions_url
+>>> dimensions_url("pub.1127419018")
+'https://app.dimensions.ai/details/publication/pub.1127419018'
 
 """
 
@@ -8,23 +14,28 @@ import click
 import time
 import json
 import sys
-import subprocess
 import os
-import re
-import webbrowser
-from itertools import islice
-from pandas import json_normalize, DataFrame
-
-from ..core.dsl_grammar import *
-
-from .html import html_template_interactive
 
 
 
 def dslquery(query_string):
-    """shortcut for running a query - meant to be used only within interactive computing environments
+    """Shortcut for running a query without instantiating dimcli.Dsl(). 
+    
+    Added for backward compatibility with legacy API tutorials. Requires file-based credentials for logging in.
+    
+    Parameters
+    ----------
+    query_string: str 
+        A valid DSL query.    
+
+    Returns
+    -------
+    DslDataset
+        A Dimcli wrapper object containing JSON data. 
+
     """
     from ..core.auth import is_logged_in
+    from ..core.api import Dsl
     if is_logged_in():
         dsl = Dsl()
         res = dsl.query(query_string, verbose=True)
@@ -32,22 +43,46 @@ def dslquery(query_string):
 
 
 def dslquery_json(query_string):
-    """shortcut for backward compatibility 
-    Same as above but returns raw JSON instead of Api.DslDataset object
+    """Shortcut for running a query without instantiating dimcli.Dsl(). Same as ``dslquery`` but returns raw JSON instead of Api.DslDataset object
+    
+    Added for backward compatibility with legacy API tutorials. Requires file-based credentials for logging in.
 
-    Pattern: `from dimcli.shortcuts import dslquery_json as dslquery`
+    Parameters
+    ----------
+    query_string: str 
+        A valid DSL query.    
+
+    Returns
+    -------
+    Dict
+        API JSON data, represented as a dict object.
+
     """
     from ..core.auth import is_logged_in
+    from ..core.api import Dsl
     if is_logged_in():
         dsl = Dsl()
         return dsl.query(query_string).json
 
 
 def dslqueryall(query_string):
-    """shortcut for running a loop query - meant to be used only within interactive computing environments
-    NOTE: this requires the file-based credentials file set up.
+    """Shortcut for running a loop query without instantiating dimcli.Dsl().
+    
+    Added for backward compatibility with legacy API tutorials. Requires file-based credentials for logging in.
+
+    Parameters
+    ----------
+    query_string: str 
+        A valid DSL query.    
+
+    Returns
+    -------
+    DslDataset
+        A Dimcli wrapper object containing JSON data.
+
     """
     from ..core.auth import is_logged_in
+    from ..core.api import Dsl
     if is_logged_in():
         dsl = Dsl()
         return dsl.query_iterative(query_string)
@@ -56,12 +91,30 @@ def dslqueryall(query_string):
 
 
 def dimensions_url(obj_id, obj_type="", verbose=True):
-    """
-    Generate a valid Dimensions URL for one of the available sources.
-    obj_id: the dimensions ID of the object
-    obj_type: one of 'publications', 'grants', 'patents', 'policy_documents', 'clinical_trials', 'researchers'
+    """Generate a valid Dimensions URL for one of the available sources.
+
+    Parameters
+    ----------
+    obj_id: str 
+        A Dimensions ID for one of the available sources.  
+    obj_type: str, optional
+        The name of the source: one of 'publications', 'grants', 'patents', 'policy_documents', 'clinical_trials', 'researchers'. If not provided, it's inferred using the ID structure.
+
+    Returns
+    -------
+    str
+        The object URL.
+
+    Example
+    ----------
+    >>> from dimcli.utils import dimensions_url
+    >>> dimensions_url("pub.1127419018")
+    'https://app.dimensions.ai/details/publication/pub.1127419018'
+
     """
     
+    from ..core.dsl_grammar import G 
+
     if obj_type and obj_type not in G.sources():
         raise ValueError("ERROR: valid sources are: " + " ".join([x for x in G.sources()]))
     if not obj_type:
@@ -76,7 +129,26 @@ def dimensions_url(obj_id, obj_type="", verbose=True):
 
 
 def dimensions_url_search(keywords_list_as_string):
-    "Returns a valid keyword search URL for Dimensions"
+    """Generate a valid keyword search URL for Dimensions.
+
+    Parameters
+    ----------
+    keywords_list_as_string: str 
+        List of search keywords.  
+
+    Returns
+    -------
+    str
+        The Dimensions URL.
+
+    Example
+    ----------
+    >>> from dimcli.utils import dimensions_url_search
+    >>> dimensions_url_search("graphene AND south korea")
+    'https://app.dimensions.ai/discover/publication?search_text=graphene%20AND%20south%20korea&search_type=kws&search_field=full_search'
+
+    """
+
     q = """https://app.dimensions.ai/discover/publication?search_text={}&search_type=kws&search_field=full_search"""
     from urllib.parse import quote   
     s = quote(keywords_list_as_string)  
@@ -86,23 +158,29 @@ def dimensions_url_search(keywords_list_as_string):
 
 
 def dsl_escape(stringa, all=False):   
-    """
-    Helper for escaping the full-text inner query string, when it includes quotes. Usage:
-
-    `search publications for "{dsl_escape(complex_q)}" return publications`
-
-    EG imagine the query string:
+    """Helper for escaping the full-text inner query strings, when they includes quotes. 
+    
+    EG with the query string:
     '"2019-nCoV" OR "COVID-19" OR "SARS-CoV-2" OR (("coronavirus"  OR "corona virus") AND (Wuhan OR China))'
     
     In Python, if you want to embed it into a DSL query, it has to become:
     '\\"2019-nCoV\\" OR \\"COVID-19\\" OR \\"SARS-CoV-2\\" OR ((\\"coronavirus\\"  OR \\"corona virus\\") AND (Wuhan OR China))'
 
-    NOTE by default only quotes as escaped. If you want to escape all special chars, pass all=True, eg
-
-    > dsl_escape('Solar cells: a new technology?', True)
-    > 'Solar cells\\: a new technology?'
-
     See also: https://docs.dimensions.ai/dsl/language.html#for-search-term
+
+    Parameters
+    ----------
+    stringa: str 
+        Full-text search component of a DSL query.
+    all: bool, default=False
+        By default only quotes as escaped. Set to True to escape all special characters (eg colons)
+
+    Example
+    ----------
+    >>> dsl_escape('Solar cells: a new technology?', True)
+    'Solar cells\\: a new technology?'
+
+    
     """
     
     if all: